+++ game/Lmd_Commands_Core.c
@@ -265,7 +265,9 @@ void Cmd_Help_f(gentity_t *ent){
 		}
 	}
 	else {
-		Disp(ent, va("^8Lugormod ^6%s ^0-- ^8by iomatix\n ^5Guide: www.github.com/iomatix/Lugormod-X\n", Lugormod_Version.string));
+		//Disp(ent, va("^8Lugormod ^6%s ^0-- ^8by iomatix\n ^5Guide: www.github.com/iomatix/Lugormod-X\n", Lugormod_Version.string));
+		trap_SendServerCommand(ent->s.number, va("chat \"^8Lugormod ^6%s ^0-- ^8by iomatix\"", Lugormod_Version.string));
+		trap_SendServerCommand(ent->s.number, "chat \"^5Guide: www.github.com/iomatix/Lugormod-X\"");
 		Cmd_Help_ListCategories(ent);
 	}
 }